+++ tests/AirplaneTest.php
@@ -17,14 +17,6 @@ class AirplaneTest extends PHPUnit_Framework_TestCase
 
     private $airplane;
 
-    //JSON of airplanes from the vault
-    private $airplanes_json;
-
-    //Array of airplanes from the vault
-    private $valid_airplanes;
-
-    //List of valid models
-    private $valid_models;
 
     //Valid first character for identifiers (S for Swallow)
     private $first_char;
@@ -32,7 +24,7 @@ class AirplaneTest extends PHPUnit_Framework_TestCase
     //Smallest allowable integer
     private $min_valid_int;
 
-    //Number of character in the longest allowable string length
+    //Number of characters in the longest allowable string length
     private $max_valid_strlen;
 
     /*
@@ -42,16 +34,10 @@ class AirplaneTest extends PHPUnit_Framework_TestCase
     {
         // Load CI instance normally
         $this->CI = &get_instance();
-        $this->airplanes_json = '[ { "id": "avanti", "manufacturer": "Piaggo", "model": "Avanti II", "price": "7195", "seats": "8", "reach": "2797", "cruise": "589", "takeoff": "994", "hourly": "977" }, { "id": "baron", "manufacturer": "Beechcraft", "model": "Baron", "price": "1350", "seats": "4", "reach": "1948", "cruise": "373", "takeoff": "701", "hourly": "340" }, { "id": "caravan", "manufacturer": "Cessna", "model": "Grand Caravan EX", "price": "2300", "seats": "14", "reach": "1689", "cruise": "340", "takeoff": "660", "hourly": "389" }, { "id": "citation", "manufacturer": "Cessna", "model": "Citation M2", "price": "3200", "seats": "7", "reach": "1550", "cruise": "748", "takeoff": "978", "hourly": "1122" }, { "id": "kingair", "manufacturer": "Beechcraft", "model": "King Air C90", "price": "3900", "seats": "12", "reach": "2446", "cruise": "500", "takeoff": "1402", "hourly": "990" }, { "id": "mustang", "manufacturer": "Cessna", "model": "Citation Mustang", "price": "2770", "seats": "4", "reach": "2130", "cruise": "630", "takeoff": "950", "hourly": "1015" }, { "id": "pc12ng", "manufacturer": "Pilatus", "model": "PC-12 NG", "price": "3300", "seats": "9", "reach": "4147", "cruise": "500", "takeoff": "450", "hourly": "727" }, { "id": "phenom100", "manufacturer": "Embraer", "model": "Phenom 100", "price": "2980", "seats": "4", "reach": "2148", "cruise": "704", "takeoff": "1036", "hourly": "926" } ]';
-        $this->valid_airplanes = json_decode($this->airplanes_json);
         $this->airplane = new Airplane();
-        $this->valid_models = [];
         $this->first_char = "S";
         $this->min_valid_int = 0;
         $this->max_valid_strlen = 64;
-        foreach ($this->valid_airplanes as $plane) {
-            $this->valid_models[] = $plane->model;
-        }
     }
 
     //TESTS FOR ID
@@ -95,7 +81,7 @@ class AirplaneTest extends PHPUnit_Framework_TestCase
             $value .= $this->first_char;
         }
         $this->airplane->id = $value;
-        $this->assertLessThanOrEqual($this->max_valid_strlen, strlen($this->airplane->id));
+        $this->assertNotEquals($value, $this->airplane->id);
     }
 
     //String characters including symbols and punctuations
@@ -103,7 +89,7 @@ class AirplaneTest extends PHPUnit_Framework_TestCase
     {
         $value = $this->first_char . "abcde!@#$%^&*()_+{}:\" <>?~`-=[];\',./\\";
         $this->airplane->id = $value;
-        $this->assertTrue(ctype_alnum($this->airplane->id));
+        $this->assertNotEquals($value, $this->airplane->id);
     }
 
     //First character is random alphanumeric character
@@ -111,7 +97,7 @@ class AirplaneTest extends PHPUnit_Framework_TestCase
     {
         $value = "x00000";
         $this->airplane->id = $value;
-        $this->assertStringStartsWith($this->first_char, $this->airplane->id);
+        $this->assertNotEquals($value, $this->airplane->id);
     }
 
     // TESTS FOR MANUFACTURER
@@ -139,7 +125,7 @@ class AirplaneTest extends PHPUnit_Framework_TestCase
             $value .= "m";
         }
         $this->airplane->manufacturer = $value;
-        $this->assertLessThanOrEqual($this->max_valid_strlen, strlen($this->airplane->manufacturer));
+        $this->assertNotEquals($value, $this->airplane->manufacturer);
     }
 
     //TESTS FOR MODEL