+++ tests/FlightTest.php
@@ -15,18 +15,28 @@ class FlightTest extends PHPUnit_Framework_TestCase
 
     private $flight;
 
+    //Valid first character for identifiers (S for Swallow)
     private $first_char;
 
+    //Longest allowable string length
     private $max_valid_strlen;
 
+    //Smallest allowable integer
+    private $min_valid_int;
+
+    //Earliest time for departure
     private $earliest_departure;
 
+    //Latest time for arrival at home base
     private $latest_arrival;
 
+    //Loading wait time between flights
     private $wait_time;
 
+    //Buffer time for getting up to cruise speed and slowing down
     private $buffer_time;
 
+    //Airport code valid string length
     private $airport_code_strlen;
 
     //Test setup.
@@ -36,11 +46,34 @@ class FlightTest extends PHPUnit_Framework_TestCase
         $this->flight               = new Flight();
         $this->first_char           = "S";
         $this->max_valid_strlen     = 64;
-        $this->earliest_departure   = "08:00";
-        $this->latest_arrival       = "22:00";
+        $this->min_valid_int        = 0;
+        $this->earliest_departure   = '08:00';
+        $this->latest_arrival       = '22:00';
         $this->wait_time            = 30;
         $this->buffer_time          = 10;
         $this->airport_code_strlen  = 3;
+
+
+
+    }
+    //TESTS FOR ID
+
+    //Valid tests for Id
+
+    //Id equal to 0
+    public function testIdValid() {
+        $value = 0;
+        $this->flight->id = $value;
+        $this->assertGreaterThanOrEqual($this->min_valid_int, $this->flight->id);
+    }
+
+    //Invalid tests for airplanes
+
+    //Id equal to less than 0
+    public function testIdInvalid() {
+        $value = -1;
+        $this->flight->id = $value;
+        $this->assertNotEquals($value, $this->flight->id);
     }
 
     //TESTS FOR CODE
@@ -84,7 +117,7 @@ class FlightTest extends PHPUnit_Framework_TestCase
             $value .= $this->first_char;
         }
         $this->flight->code = $value;
-        $this->assertLessThanOrEqual($this->max_valid_strlen, strlen($this->flight->code));
+        $this->assertNotEquals($value, strlen($this->flight->code));
     }
 
     //String characters including symbols and punctuation
@@ -100,17 +133,36 @@ class FlightTest extends PHPUnit_Framework_TestCase
     {
         $value = "x00000";
         $this->flight->code = $value;
-        $this->assertNotEquals($this->flight->code, $value);
+        $this->assertNotEquals($value, $this->flight->code);
     }
 
     //TESTS FOR AIRPLANES
-    /*
-     *
-     * TO BE IMPLEMENTED
-     *
-     */
+
+    //Valid tests for airplanes
+
+    //Airplanes (ID) equal to 0
+    public function testAirplanesValid() {
+        $value = 0;
+        $this->flight->airplanes = $value;
+        $this->assertGreaterThanOrEqual($this->min_valid_int, $this->flight->airplanes);
+    }
+
+    //Invalid tests for airplanes
+
+    //Airplanes (ID) equal to less than 0
+    public function testAirplanesInvalid() {
+        $value = -1;
+        $this->flight->airplanes = $value;
+        $this->assertNotEquals($value, $this->flight->airplanes);
+    }
 
     //TESTS FOR DEPARTAIRPORT
+
+    public function testDepartAirportType() {
+        $value = "XYZ";
+        $this->flight->departAirport = $value;
+        $this->assertInternalType('string', $this->flight->arriveAirport);
+    }
 
     //Valid tests for departAirport
 
@@ -153,9 +205,59 @@ class FlightTest extends PHPUnit_Framework_TestCase
         $this->assertNotEquals($this->flight->departAirport, $value);
     }
 
+    //TESTS FOR DEPARTTIME
+
+    public function testDepartTimeType() {
+        $value = '08:00';
+        $this->flight->departTime = $value;
+        $this->assertInternalType('string', $this->flight->departTime);
+    }
+
+    //Valid tests for departTime
+
+    //String is earliest departure time
+    public function testDepartTimeEarliestValid() {
+        $value = '08:00';
+        $this->flight->departTime = $value;
+        $this->assertGreaterThanOrEqual(new DateTime($this->flight->departTime), new DateTime($this->earliest_departure));
+    }
+
+    //Departure time string is earlier than arrival time
+    public function testDepartTimeEarlierThanArriveTimeValid() {
+        $value1 = '22:00';
+        $this->flight->arriveTime = $value1;
+        $value2 = '21:59';
+        $this->flight->departTime = $value2;
+        $this->assertLessThan(new DateTime($this->flight->departTime), new DateTime($this->flight->arriveTime));
+    }
+
+    //Invalid tests for departTime
+
+    //String is earlier than earliest depature time
+    public function testDepartTimeEarliestInvalid() {
+        $value = '7:59';
+        $this->flight->departTime = $value;
+        $this->assertNotEquals($this->flight->departTime, $value);
+    }
+
+    //Departure string is later than arrival time
+    public function testDepartTimeLaterThanArriveTimeInvalid() {
+        $value1 = '21:59';
+        $this->flight->arriveTime = $value1;
+        $value2 = '22:00';
+        $this->flight->departTime = $value2;
+        $this->assertNotEquals($value2, $this->flight->departTime);
+    }
+
     //TESTS FOR ARRIVEAIRPORT
 
-    //Valid tests for departAirport
+    public function testArriveAirportType() {
+        $value = "XYZ";
+        $this->flight->arriveAirport = $value;
+        $this->assertInternalType('string', $this->flight->arriveAirport);
+    }
+
+    //Valid tests for arriveAirport
 
     //String length equal to valid number of chars
     public function testArriveAirportStrlenValid()
@@ -194,5 +296,49 @@ class FlightTest extends PHPUnit_Framework_TestCase
         $value = "xYz";
         $this->flight->arriveAirport = $value;
         $this->assertNotEquals($this->flight->arriveAirport, $value);
+    }   
+    
+    //TESTS FOR ARRIVETIME
+
+    public function testArriveTimeType() {
+        $value = '22:00';
+        $this->flight->arriveTime = $value;
+        $this->assertInternalType('string', $this->flight->arriveTime);
+    }
+
+    //Valid tests for departTime
+
+    //String is earliest departure time
+    public function testArriveTimeLastestValid() {
+        $value = '22:00';
+        $this->flight->arriveTime = $value;
+        $this->assertLessThanOrEqual(new DateTime($this->latest_arrival), new DateTime($this->flight->departTime));
+    }
+
+    //Arrival time string is later than arrival time
+    public function testArriveTimeLaterThanDepartTimeValid() {
+        $value1 = '21:59';
+        $this->flight->departTime = $value1;
+        $value2 = '22:00';
+        $this->flight->arriveTime = $value2;
+        $this->assertGreaterThan(new DateTime($this->flight->departTime), new DateTime($this->flight->arriveTime));
+    }
+
+    //Invalid tests for departTime
+
+    //String is earlier than earliest depature time
+    public function testArriveTimeLatestInvalid() {
+        $value = '22:01';
+        $this->flight->arriveTime = $value;
+        $this->assertNotEquals($value, $this->flight->arrivetTime);
+    }
+
+    //Departure string is later than arrival time
+    public function testArriveTimeEarlierThanDepartTimeInvalid() {
+        $value1 = '22:00';
+        $this->flight->departTime = $value1;
+        $value2 = '21:59';
+        $this->flight->arriveTime = $value2;
+        $this->assertNotEquals($value2, $this->flight->arriveTime, "arriveTime is set with invalid value");
     }
 }